TagController update
Got rid of some redundant code and set the nametag scaling on the prefab
