+++ Assets/Custom/Scripts/StaminaTagController.cs
@@ -1,27 +1,10 @@
-﻿using UnityEngine;
-using UnityEngine.UI;
+﻿using UnityEngine.UI;
 
 namespace com.terranovita.botretreat
 {
-    public class StaminaTagController : MonoBehaviour, IBotDependant
+    public class StaminaTagController : TagController, IBotDependant
     {
-
-        public GameObject BotGameObject { get; set; }
-
-        void Start()
-        {
-
-        }
-
-        void Update()
-        {
-            // Position the name tag centered above the bot.
-            var botPosition = BotGameObject.transform.position;
-            transform.position = new Vector3(botPosition.x, 2.3f, botPosition.z);
-            // Always let the name tags look directly at the camera.
-            var mainCameraRotation = Camera.main.transform.rotation;
-            transform.LookAt(transform.position + mainCameraRotation * Vector3.forward, mainCameraRotation * Vector3.up);
-        }
+        public StaminaTagController() : base(1.8f) { }
 
         public void UpdateBot(Bot bot)
         {
