+++ Assets/Custom/Scripts/TagController.cs.meta
@@ -1,0 +1,12 @@
+fileFormatVersion: 2
+guid: 594c7be5712e00845b2b86e9de6d81ab
+timeCreated: 1455202567
+licenseType: Free
+MonoImporter:
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
